Added some styling to the main.handlebars. Server should start up now.
