+++ db/schema.sql
@@ -1,2 +1,7 @@
+<<<<<<< Updated upstream
+=======
+DROP DATABASE IF EXISTS user_db;
+CREATE DATABASE user_db;
+>>>>>>> Stashed changes
 DROP DATABASE IF EXISTS salon_db;
 CREATE DATABASE salon_db;
