+++ public/css/style.css
@@ -1,0 +1,199 @@
+body{
+background-color: beige;
+}
+
+.banner {
+    background-color: rgb(228, 212, 183) !important;
+    background-image: url(../images/Salonbanner.png);
+    background-repeat: no-repeat;
+    background-size: cover;
+    width: 100%;
+    margin: 0px;
+    padding: 0px;
+    height:15vh;
+}
+header{
+    display: flex;
+    flex-direction: column;
+}
+
+#smallScreenButton{
+    display: none;
+}
+
+a {
+    text-decoration: none !important;
+}
+nav {
+    background: rgb(207,201,176);
+background: linear-gradient(90deg, rgba(207,201,176,1) 0%, rgba(182,175,146,1) 49%, rgba(207,201,176,1) 100%);
+    width: 100%;
+    padding: 2px;
+    flex-shrink: 0;
+    z-index: 2;
+}
+aside {
+    z-index: 0;
+}
+main{
+    flex-grow: 1;
+    z-index: 0;
+}
+
+.unadornedAnchor {
+    text-decoration: none !important;
+    color:black;
+}
+a:hover{
+    color:rgb(15, 15, 15)
+}
+.centered {
+    width: 100%;
+    justify-self: center;
+    justify-content: center;
+    align-items: center;
+    align-content: center;
+}
+.pl-5{
+    padding-left: 15px;
+}
+.mb-5{
+    padding-bottom: 20px;
+}
+.mt-5{
+    margin-top: 20px;
+}
+
+.navbutton{
+    width:150px;
+    height:35px;
+    border-top: 0;
+    border-bottom: 0;
+    border-left: 1px solid #857e64;
+    border-right: 1px solid #857e64;
+    color:#857e64;
+    vertical-align: middle;
+    font-weight: 200;
+    font-size: 25px;
+    padding: 2px 0;
+}
+.logbutton{
+    display: block;
+    width:50px;
+    height:35px;
+    border: 1px solid rgb(100, 93, 70);
+    background: rgb(203,196,170);
+background: linear-gradient(175deg, rgba(203,196,170,1) 0%, rgba(164,157,131,1) 49%, rgba(133,126,100,1) 100%);
+}
+.yarr {
+    background-color: white;
+    color:black;
+    border: 1px solid black;
+    margin: 20px;
+    width: 120px;
+    height: 600px;
+}
+.mainsection{
+    width: 80%;
+}
+.navigation{
+    justify-content: center;
+    flex-wrap: nowrap;
+    text-align: center;
+}
+#logo{
+    justify-content: center;
+    padding:15px;
+}
+#logocard{
+    background-color: rgb(228, 218, 133);
+    padding: 10px;
+    border-radius: 25px;
+}
+
+@media screen and (max-width:1700px){
+    #logo {
+        padding-top:5px
+    }
+    nav {
+        margin-top: 2px;
+    }
+    .mainsection{
+        width: 75%;
+    }
+}
+/* Adjustments for screens up to 768px wide */
+@media screen and (max-width: 768px) {
+    nav {
+        width: 100%;
+        padding: 2px;
+        margin-top: 10px;
+        padding-top: 0;
+    }
+    .yarr{
+        display: none;
+    }
+    #logo{
+        margin: 0;
+        padding-top: 10px;
+    }
+    #logocard{
+        padding: 5px;
+        margin:0px;
+    }
+  }
+  
+  /* Adjustments for screens up to 480px wide */
+  @media screen and (max-width: 480px) {
+    nav {
+        margin-top: 15px;
+    }
+    #logo {
+        padding-top: 15px;
+    }
+    #wideScreenButtons{
+        display: none;
+    }
+    #smallScreenButton{
+        display:block;
+    }
+    
+    .navbutton {
+      width: 100px;
+      height: 25px;
+      font-size: 0.8rem;
+    }
+    .logbutton {
+      width: 35px;
+      height: 25px;
+      font-size: 0.7rem;
+    }
+    .yarr {
+      width: 95%;
+      margin: 10px auto;
+    }
+    .mainsection {
+      width: 95%;
+      margin: 10px auto;
+    }
+  }
+  
+  /* Adjustments for screens up to 360px wide */
+  @media screen and (max-width: 360px) {
+    nav {
+        margin-top: 20px;
+    }
+    .banner {
+      border-bottom: 0;
+    }
+    .navbutton {
+      width: 90px;
+      height: 20px;
+      font-size: 0.7rem;
+    }
+    .logbutton {
+      width: 30px;
+      height: 20px;
+      font-size: 0.6rem;
+    }
+  }
