+++ public/js/register.js
@@ -1,0 +1,29 @@
+const registerFormHandler = async (event) => {
+    // Stop the browser from submitting the form so we can do so with JavaScript
+    event.preventDefault();
+  
+    // Gather the data from the form elements on the page
+    const username = document.querySelector('#user-register')
+    const email = document.querySelector('#email-register').value.trim();
+    const password = document.querySelector('#password-register').value.trim();
+  
+    if (email && password) {
+      // Send the e-mail and password to the server
+      const response = await fetch('/api/users/register', {
+        method: 'POST',
+        body: JSON.stringify({user, email, password }),
+        headers: { 'Content-Type': 'application/json' },
+      });
+  
+      if (response.ok) {
+        document.location.replace('/');
+      } else {
+        alert('Failed to register new user.');
+      }
+    }
+  };
+  
+  document
+    .querySelector('.register-form')
+    .addEventListener('submit', registerFormHandler);
+
