+++ views/register.handlebars
@@ -1,0 +1,21 @@
+<h2>Register</h2>
+
+<form class="form register-form">
+    <div class="register-group">
+    <label for="user-register">Username:</label>
+    <input class="form-input" type="text" id="email-register" />
+  </div>
+  <div class="form-group">
+    <label for="email-register">email:</label>
+    <input class="form-input" type="text" id="email-register" />
+  </div>
+  <div class="form-group">
+    <label for="password-register">password:</label>
+    <input class="form-input" type="password" id="password-register" />
+  </div>
+  <div class="form-group">
+    <button class="btn btn-primary" type="submit">register</button>
+  </div>
+</form>
+
+<script src="/js/register.js"></script>
